training-sidecar/pyproject.toml
[project]
name = "img-tagger-sidecar"
version = "0.1.0"
description = "Python sidecar for img-tagger — handles LoRA training orchestration and VLM captioning"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn>=0.34,<1",
    "websockets>=14,<15",
    "pydantic>=2.10,<3",
    "pyyaml>=6,<7",
    # VLM captioning backend — install on demand via the "vlm" extra once
    # we're ready to swap the mock provider. Keeping it optional means the
    # sidecar can run without a CUDA toolchain.
]

[project.optional-dependencies]
vlm = [
    # Install with: uv sync --extra vlm
    # Default: CPU wheels from abetlen's wheel index (cross-platform).
    # For Windows CUDA, you need to build from source:
    #   $env:CMAKE_ARGS = "-DGGML_CUDA=on"
    #   uv pip install --force-reinstall --no-cache-dir llama-cpp-python
    "llama-cpp-python>=0.3.0,<1",
]

[tool.uv]
package = false

[tool.uv.sources]
llama-cpp-python = { index = "llama-cpp-cpu" }

[[tool.uv.index]]
name = "llama-cpp-cpu"
url = "https://abetlen.github.io/llama-cpp-python/whl/cpu"
explicit = true
